usbc: Add set and get functions for USB-C port count
Add set and get functions to retrieve the number of usbc
ports as well as to set the number of ports supported for
specific boards at runtime. Change references for
Config - CONFIG_USB_PD_PORT_MAX_COUNT to get function
-board_get_usb_pd_port_count(). Also, enable a new config
CONFIG_USB_PD_COUNT_RUNTIME for the purpose of setting
and getting the number of ports at runtime. In future,
the overridable function to retrieve port number can be
removed.

BUG=b:194163586
BRANCH=none
TEST=make buildall -j

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: I568abff0a3e1e8aed66a298e6c1dc355ffbbc138