charge_manager:Overridable function to get charge port count
Currently, config- CHARGE_PORT_COUNT decides the count of ports
that can be used for charging.Intention of this CL is to have generic
overridable functions, so that number of ports, which can be used for
charging can be decided at run time based on the type of boards. For ex:
Number of ports that can be used for charging for M & N are configured at
run time.There are totally 3 ports for M/N for charging, including 2
typec ports and a charger AC port. This CL can help in retrieving the
charge port count at run time.

BUG=b:194163586
BRANCH=none
TEST=make build all

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: I67ea6b566e149b78ac4521538257c2c362834c88