zephyr: cros_cbi bug fix and test
The cros,cbi driver was found to contain a bug. During
CBI_SSFC_INIT_DEFAULT_ID() the static cached_ssfc variable was being
initialized. While all the functions using the ssfc values were using
the ones stored in the device's data struct. This means that the case
statements generated were always returning 0 for every case.

Additionally, the call to cros_cbi_ssfc_get_parent_field_value() was
returning 0 for error which is also a valid value. This was also fixed
so it is possible to distinguish between an error and a value.

Tests were added to check all code paths.

BRANCH=none
BUG=b:202789410
TEST=zmake configure --test zephyr/test/drivers

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: Icac1491c72e95e4700ce696f91f2cfaf56e0ebff