ap_ro_verification: Do not allow over USB
It should not be possible to trigger AP RO verification by sending the
vendor command over the USB interface, it should be triggered only
when the user enters the existing key sequence, or when coming from
the AP as a TPM command.

When verification is triggered by the key sequence the verification
function is still invoked as a TPM vendor command, because a context
switch into TPM task is required to be able to use its large stack.
The problem is that when TPM task processes context switching
commands, it unconditionally sets the bit indicating that they come
from USB. This is required for some CCD commands, but not for the AP
RO verification request.

As a fix, let's not set the 'FROM_USB' flag for the AP RO verification
command when it is processed by the TPM task.

BUG=b:202902506
TEST=AP RO verification still could be triggered by key presses and
     through a TPM command coming from the AP, but is rejected if sent
     over USB by the host.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: Ib3b5bd9867cc3038d9123b23c7a25886331a179c