firmware_builder: Call build_with_clang.py script
The util/build_with_clang.py script helps us validate that boards that
build successfully with clang continue to do so. As compilation errors
are fixed for boards, they can be added to the list.

The script is called from firmware_builder.py as part of the CQ process.

BRANCH=none
BUG=b:172020503
TEST=CQ passes

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I8d9384e04dbfc25191d5ebf93425e6d178631168