adlrvp: Retrieve dedicated charger ports num at runtime
The number of dedicated charger ports in ADL is based
on the number of  typec based ports present. The num of typec
ports for ADL-P/M/N is retrieved at run time based on
board id. Hence, the overridable function to get the
number of dedicated charge ports must return the number of
typec usb ports.

BUG=b:194163586
BRANCH=none
TEST= test on ADL

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: I5188cd25bfa9fafccfaaf33dbc2cedaca87e670d