i2cp: reduce the number of i2c flog errors
Cap the number of I2C flog errors at 2 per boot. The timestamps of the
two events can indicate if the unwedge issues happen successively or if
they happen after a long period of time. The successive logs are likely
benign. If we see individual i2c errors, we can investigate those
issues.

BUG=b:146067724
TEST=check the number of i2c log messages on coral in s0ix

Change-Id: I3f04c85e6233bf0c790db0d40a85aab3c927b9b8
Signed-off-by: Mary Ruthven <mruthven@chromium.org>