core/cortex-m: Use original __data_lma_start when compiling with gcc
https://crrev.com/c/3169964 changed __data_lma_start from "." to
LOADADDR(.data) to fix dartmonkey when compiling with clang. This
changed the LMA address on some boards, such as voxel. Even though the
new address looks correct (matches start of .data section), we are
seeing an exception on voxel during startup. For now, keep clang and gcc
separate until we can diagnose the issue.

BRANCH=none
BUG=b:172020503, b:202863352
TEST=make buildall -j

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I8313210c89559a6159f311821a18e7b310a3ada6