ap_ro_verification: do not stop on failing FMAPs
To prevent denial of service attack when a fake FMAP structure is
placed somewhere in the AP flash, then detected by the GSC and
rejected as corrupted, do not stop after finding an FMAP which
includes a pointer to a GVD which fails to verify.

This means the entire flash needs to be scanned, so this patch
eliminates the approach where the flash is scanned at decreasing
intervals until an FMAP section is found. Check all locations at 4K
aligned addresses instead and keep looking until a valid GVD is
located or the entire flash is scanned.

BUG=b:141191727
TEST=created a fake FMAP entry placed into the RW_A space of a
     guybrush image, with a corrupted GVD, programmed the modified
     image on a guybrush and attempted AP RO verification.

     Observed the GSC report the inconsistent GVD contents and then
     find the proper GVD structure in a different FMAP structure, and
     successfully validate the GVD structure.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: Ic8a930af63e1b90343d8cae6a86e65b06decebfb