TCPMv2: Only clear discovery data on disconnect
Currently, the HOOK_USB_PD_DISCONNECT hook runs before mode exit in the
TC.  This means that if the PE clears the mode data, we won't know to
exit active modes.

Decouple this dependency by leaving the alt mode exit code in charge of
clearing the modes on disconnect.

BRANCH=None
BUG=None
TEST=on voxel, disconnect DP monitor and verify HPD is set low

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Iebfddf4b7880dcfbeab08bdca426175bc549c833