zephyr: drivers: Add PS8xxx unit tests
This patch adds unit tests for code of PS8xxx driver that is used
for all devices.

PS8805 is first device from PS8xxx family which will be tested.
All devices that are used for USBC port 1 are moved to i2c1 emulated
bus, because addresses required for PS8805 are not available on i2c0
bus.

BUG=b:184857030
BRANCH=none
TEST=make configure --test zephyr/test/drivers

Signed-off-by: Tomasz Michalec <tm@semihalf.com>
Change-Id: I0f146847c038014258e517c82b3268f42079efc9