board_id: fix an incorrect offset calculation
The INFO_BOARD_ID_OFFSET value includes the offset of the board ID
field, there is no need to add the field offset again.

This incorrect calculation is not a problem, because the offset of the
'bid' field is zero, but the logic is wrong.

BUG=none
TEST=board ID value is still reported properly.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I3e7061a930d751d2cf13113b1e519e8f976195bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3236754
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
(cherry picked from commit 2b948afb28d0674eebb4dc0ece088fbf95f6a142)