zephyr: emul: Fix bad type cast in emul_sn5s330
Due to a bad pointer cast in the sn5s330_emul_peek_reg function, it
would return garbage values from memory. This was originally passing
because the test was only testing if two bits were set which just so
happened to be the case given the memory layout when it was written.

BRANCH=none
BUG=b:203364783
TEST=zmake configure --test zephyr/test/drivers

Signed-off-by: Aaron Massey <aaronmassey@google.com>
Change-Id: I41f3e8316d879b764c41d27dfacdb05a753faf9f