common/acpi: Disable clang warning
When building with clang, it warns:

common/acpi.c:138:11: error: result of comparison of constant 256 with
expression of type 'uint8_t' (aka 'unsigned char') is always false
[-Werror,-Wtautological-constant-out-of-range-compare]
            addr >= EC_ACPI_MEM_MAPPED_BEGIN + EC_ACPI_MEM_MAPPED_SIZE) {
            ~~~~ ^  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

It doesn't hurt to leave the check as-is. The values of
EC_ACPI_MEM_MAPPED_BEGIN + EC_ACPI_MEM_MAPPED_SIZE may not be greater than or
equal to 255 on every board.

BRANCH=none
BUG=b:172020503
TEST=CC=clang make BOARD=voema

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: Ibb27421957e99f2f636331fbaa51dab6fa768744