Guybrush: Handle dead battery boot in more scenarios
When booting with no battery and multiple chargers (ex. suzy-q), we may
end up in scenarios where we have more than one TCPC in dead battery
mode.  Attempt to get these TCPCs reset whenever we can, and always err
on the side of selecting the dead battery port as charge port.

This also corrects a typo in the case of no charge port selected.

BRANCH=None
BUG=b:183660105
TEST=on guybrush:
- boot with cutoff battery at 0% and two chargers, verify one is reset
and the other is selected.  Re-plug charger that isn't selected and
verify it still isn't selected,  Charge to 4% and verify new charger can
be selected and old port is reset.
- boot with cutoff battery at 9% and one charger, verify dead battery
port is immediately reset

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I4d4ebf314efbbcde99c820a78f5bd4639d08e0a5