common/firmware_image.lds.S: Add empty text section
Create empty text section. Otherwise clang/lld will create one and put
it at an address it selects. Even though the section is empty, if the
LMA is beyond the end of the flash, ec.bin will be too large since the
empty space is filled with 0xFF.

BRANCH=none
BUG=b:172020503
TEST=./util/compare_build.sh -b all -j 120
     => MATCH

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I48f8019fe6b77f880de0d1a0a4a176b3506f3579