TCPMv2: avoid Hard reset in S5/G3 state when receive NOT_SUPPORT
In this change CL:/3161320, when system state in S5/G3 state EC will
send the Hard reset and cause AC only can't power on.
This is because EC force to be vconn source when adapter response not
support vconn swap.

We add the new flag to record the port partner response not support
vconn swap command and don't send the Hard reset when this flag is
set.

BUG=none
BRANCH=dedede
TEST=make BOARD=cret, DUT can power on with AC only

Signed-off-by: Josh Tsai <josh_tsai@compal.corp-partner.google.com>
Change-Id: Ic29964967b7ec43f9ea6f17b4c36a65cbaaa4c3e