zephyr: ppc: Add support for alternate chips
Create an additional ppc_chips_alt array to store info about alternative
ppc chips that can be used depending on some conditions e.g. a board
version.

Such alternative chips should be added to a dts file with the
"alternate-for"cproperty to point to a ppc chip that should be replaced.

BUG=b:194432779
TEST=zmake testall & Make sure PD works on Herobrine
BRANCH=main

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: Ib61ac767e7377c87c686fde9c0d765ffe682e582