core/cortex-m0: fix vecttable.c when compiling with clang
This is the same as commit 0907090490b07357902927f089228d63a775cdf0,
which made the same change for core/cortex-m/vecttable.c.

Text section name is adding ",\"a\" @" to the section name.

arm-none-eabi-objdump  --disassemble-all \
./build/nocturne_fp/RO/core/cortex-m/vecttable.o

Before:
Disassembly of section .text.vecttable,"a" @:

00000000 <vectors>:
    ...

After:
Disassembly of section .text.vecttable:

00000000 <vectors>:
    ...

Comparing the text.vecttable elf section headers for gcc before and
after this change, there is no difference to flags:

arm-none-eabi-objdump -h \
./build/nocturne_fp/RO/core/cortex-m/vecttable.o

.text.vecttable 00000298  00000000  00000000  00000050  2**2
	     CONTENTS, ALLOC, LOAD, RELOC, READONLY, CODE

One difference when compiling with clang is that it sets the DATA
attribute in the header, while gcc sets CODE. Since this is an array of
addresses, not executable code, I think clang is actually correct:

.text.vecttable 00000298  00000000  00000000  00000060  2**2
	      CONTENTS, ALLOC, LOAD, RELOC, READONLY, DATA

BRANCH=none
BUG=b:172020503
TEST=none

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I8d84f44266fe3bfe3a942401d9fdbe7760c6af53