tcpm: Add port flag indicating if the TCPC controls VCONN
This adds a port flag to indicate that its TCPC needs to control VCONN.
We have a global compile time flag (CONFIG_USB_PD_TCPC_VCONN) to enable
TCPC control of VCONN. However, on some device architectures, such a
global setting is not suitable for all ports in the system. So, add a
port flag to enable VCONN control where it would otherwise be disabled.

BRANCH=none
BUG=b:204102039
TEST=with reset of patch stack, external monitor links up using active
	TBT4 cable

Change-Id: I9028f65a8b10b2f8eb5b614a256d90e05064b4c3
Signed-off-by: Caveh Jalali <caveh@chromium.org>