kodama: drop kodama board
Kodama is the tightest space board in kukui family, and we
have to constanly disable configs to free up space.
Since we don't really need the board in main branch as the
firmware branch has been created for a long time, just drop
this board on main branch.

BUG=none
TEST=makd buildall
BRANCH=main

Change-Id: I7f958ad7b97c191a8433d82987f8876add8c5dc3
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>