bq25710: Add PSYS sensing configuration
This adds a new config option to enable the PSYS monitoring feature on
the BQ25710, BQ25720 chargers. The register definition for configuring
PSYS on the BQ25720 is expanded to 2 bits, giving more control to meaure
battery and AC separately. We keep it simple and include both when
measurement is enabled.

BRANCH=none
BUG=b:195615830
TEST=with rest of patch series, psys values reported by
	dump_intel_rapl_consumption look reasonable

Change-Id: I0f299c6a24d20ef5bdcda13de74b30ba0c2d5d3c
Signed-off-by: Caveh Jalali <caveh@chromium.org>