chip/stm32/usb_spi.c: Fix bug crippling full duplex
It seems that the logical condition of an ifdef had been accidentally
negated.  Furthermore, in cases where full duplex support is present,
the control flow should reach the final else branch of the if
statement, which it previously would not when a full duplex
transaction was requested.  These two mistakes makes it such that SPI
full duplex forwarding from USB cannot work on e.g. servo v4 (and on
the future HyperDebug board that I am working on.)  I have tested this
change as part of my HyperDebug code under development, as well as
tested that the change compiles on servo_micro.  I have not attempted
full duplex transactions on servo micro, as I do not currently have a
setup that I can use to easily verify.  Looking at the code, one can
convince oneself that behavior can only change in cases where the code
would previously either reject a transaction explicitly, or silently
drop it resulting in USB timeout.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>

BUG=b:192262089
BRANCH=None
TEST=make BOARD=servo_micro

Change-Id: I5767d91c5b1ada4e54eec7ecd0512345448288e7