zephyr: zmake: Upgrade to new BUILD.py format
go/zephyr-variants defines a BUILD.py format that replaces
zmake.yaml.  BUILD.py is inspired by Bazel build definitions, allowing
for arbitrary Python code to be used as the config format.  This
enables defining multiple projects in a single directory, and projects
to make program-specific customizations to the config definition
functions.

The other following changes are of note:

- Projects can still be referenced by directory name, but can be
  referenced by the much shorter project name.  For example, both of
  the following work equivalently:

  $ zmake configure -b zephyr/projects/volteer/delbin
  $ zmake configure -b delbin

- The default build directory was changed to:
  build/zephyr/${name}

This implements the new format, drops the old YAML+jsonschema, and
migrates all our projects to the new format.

BUG=b:193815337
BRANCH=none
TEST=unit tests
TEST=zmake testall

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: Ib1e53d9abf558063117cc3f86a5d636e69cf77fe