zephyr: zmake: Rename project name field
As discussed rename the project name field to "project_name" and make
it a required as a keyword argument.

For tests, the name project_name makes a bit less sense, so call it
test_name, automatically prepend "test-" to all tests, and allow it to
be specified either as a keyword or positional argument.

BUG=b:193815337
BRANCH=none
TEST=unit tests pass

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: Idd1212da98382ddd9961da0fda4078595f540fc8