npcx: i2c: adjust HLDT for better I2C timing at 100 KHz
It was reported that the timing Tsu:sta (i.e. the setup time of repeat
start) is out of spec or on the margin when I2C runs at 100 KHz. Per
our simulation, we suggest to adjusts the HLDT value to get a better
timing.
			      | Before	   | After
	-----------------------------------------------
	I2C src clock = 15 MHz| HLDT = 0x7 | HLDT = 0x23
	I2C src clock = 20 MHz| HLDT = 0x7 | HLDT = 0x2A

BRANCH=none
BUG=b:204233183
TEST=pass "make buildall"
TEST="i2cxfer r 2 0x48 0x3" on npcx9 EVB; check the setup time of repeat
start becomes longer.
TEST=stress test with "i2cxfer r 5 0x28 0 1" over 500 times, no symptom
occured on Volteer. (Port 5 is configured to use 100 Khz.)

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: I44a41bb7e029ea3467d36789d7ca6f9144dfc577