bq25710: [CO4] Allow VSYS_UVP to be configured
This adds support for customizing the VSYS under voltage (VSYS_UVP)
lockout threshold. This is a 3 bit field with default value 0. The
actual voltage encoded is (0.8 * <value> + 2.4), allowing a range 0f 2.4
V to 8.0 V to be specified.

The CONFIG_BQ25720_CHARGE_OPTION_4_VSYS_UVP_CUSTOM option specifies if
this field should be set. If so, it is set to the value of
CONFIG_BQ25720_CHARGE_OPTION_4_VSYS_UVP.

BRANCH=none
BUG=b:185190976
TEST=buildall passes

Signed-off-by: Caveh Jalali <caveh@chromium.org>
Change-Id: Idc93558bfcaf394eb48b967a175e85cbb05ca10a