bq25710: Set IL_AVG to 10A on bq25720
This reduces the converter inductor average current clamp to 10 A on the
bq25720 from the default 15 A. This means both chip variants will use
the same 10 A limit.

All current designs using the bq25720 prefer the 10 A setting, so it
makes sense to over-ride the default on the bq25720 to mimic the bq25710
behavior. There is no need to make this configurable at this time.

BRANCH=none
BUG=b:185190976,b:192108571
TEST=buildall passes

Signed-off-by: Caveh Jalali <caveh@chromium.org>
Change-Id: I8d903eb8d46e145467b26fe8d6f6336ca20f3f5a