PoC: Introduce S4 as a real power state
In order to support hibernate (suspend to disk) on some systems, there
are normally two choices for hibernate's destination power state: shutdown
and S4. On most systems, shutdown is the logical choice for Chrome OS,
since the wake sources are the same, and device state is properly
saved/restored across an S5/G3 transition.

However on Brya devices with Intel Keylocker technology, there is an IWKey
(intermediate wrapping key) which software by design is not allowed to read.
Intel, being no stranger to the concept of hibernate, provisioned
support to save and restore this register in a platform area, while
still keeping its contents inaccessible to software. However,
architecturally they only guarantee this mechanism works down to S3/S4.

This means in order to preserve keylocker contents during hibernation,
shutdown is no longer an option. We must camp out in the architecturally
designated S4 state during hibernation on devices with Intel Keylocker.
The EC has long ignored this as a state since the OS doesn't support
entering it. This needs to change.

This patch is a working proof of concept for entering and staying in S4
for hibernate. It attempts to treat S4 similarly to how S3 is treated: a
power state we transition through most of the time to come up and down,
but also a suspend state we can camp out in for long periods. I can
hibernate on a volteer board and stay in S4 until a power button press
wakes me back up to resume.

This patch isn't ready to submit for a number of reasons:
 * It removes POWER_S3S5 and POWER_S5S3, to force me to find every place
   where we need to insert S4 into the sequence. This breaks most other
   boards, as I've only patched up what's needed for volteer. We'd need
   to go through and decide if S5 <-> S3 directly is a valid state
   transition, or if we should truly always be bouncing through S4.
 * It doesn't add the conditional compilation for S4, if that's
   something we want, for boards that don't even have a SLP_S4 pin.
 * It unconditionally starts looking at SLP_S5, where we used to watch
   SLP_S4 to jump between the S5 <-> S3 states. Is this safe to do?
 * There's probably zephyr considerations.
 * Tests?

Signed-off-by: Evan Green <evgreen@chromium.org>

BRANCH=None
BUG=b:204947672
TEST=hiberman hibernate --test-keys on volteer

Change-Id: Icf4798fa517d40ad652a278bbea2051e4c9fb118