bq25710: Refactor prochot option 1 register initialization
This separates the initialization of the prochot option 1 register into
a dedicated function. Multiple bit-fields will need to set in this
register based on board configuration, so it's best to group these
operations in one place.

Also, numeric/boolean values are replaced with symbolic values when
setting register fields for clarity where appropriate.

BRANCH=none
BUG=b:185190976
TEST=buildall passes

Signed-off-by: Caveh Jalali <caveh@chromium.org>
Change-Id: I73e6bd4f719f19ae21677b9fca90f10cf05665c1