zephyr: chip/it8xxx2: add power policy
This power policy defines that if deep sleep is allowed and
an interval of five seconds, the system can enter the suspend
state.

The UART driver uses pm_constraint_set counter to indicate that
UART Tx is in use. And use constraint_get in the power policy
to get the counter determine if the system can enter the
suspend state.

BUG=b:192354255
BRANCH=none
TEST=zmake testall
And test on hayato. If CONFIG_PM, CONFIG_DEVICE, CONFIG_PM_POLICY_APP
are enabled, the console messages normally before system enters suspend.

Signed-off-by: Tim Lin <tim2.lin@ite.corp-partner.google.com>
Change-Id: I25a9de8202c84fbb069fe66e95ce0415b2b87470