zephyr: Configure tcpci gpio alerts in test
Add GPIOs to overlay.dts.

Setup GPIO interrupts in gpio_map.h.

Implement tcpc_alert_event. This should probably be in a zephyr stub
instead, but today, this is board specific.

Add init hook to enable the usbc interrupts.

Interrupts have to be enabled by a driver or board, and these are
normally enabled by the board.

BRANCH=None
BUG=b:201314530
TEST=Observed in gdb.

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: I23bb328ffd5402e923abf8464395ca1dd2dd86de