zephyr: flash: npcx: Disable flash QE bit at init
In NPCX9 production devices, the flash status register’s quad enable
bit (non-volatile) will be set by default. When the QE bit of Status
Register-2 is set for Quad I/O, the WP# signal is not supported since
this pin is used as IO2. This CL clears the QE bit at init to make sure
WP# support is enabled.
This change aligns with the change in the CL:3246881 of ECOS.

BRANCH=none
BUG=none
TEST=Set QE bit & reboot the ec. Check the QE bit is disabled.

Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
Change-Id: Ide93ce466b61ed3358488bb91c07f9a62bce7fa1