zephyr: Drop support for kohaku
For context: this build was created so that we had a device that could
be publicly purchased to play with Zephyr on in January 2021.  With a
little passage of time, we now have lazor, limozeen, and delbin as
well.

This build supports nothing other than power sequence to S0 (no
keyboard, battery, charging, etc.) so sadly it's hardly usable for
anything at the moment.  Delete it so it does not create confusing
directory structure for how we want variants to be created.

BUG=b:193814903
BRANCH=none
TEST=zmake testall

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: Ic0fa3099c19a484513b23c7826376604cf8d2f22