intel_x86: Apply chipset resume init and suspend complete hooks
The CL:2321875 introduced a hook for chipset resume init and
suspend complete. The application was applied for SPI driver for
sc7180.

This patch applies to intel_x86 as well. The benefit that we can
initial the devices (such as PCHG) prior the chipset resume and
shutdown the devices after the chipset suspend is completed.

BUG=b:205675485
BRANCH=None
TEST=make buildall -j

Signed-off-by: Devin Lu <Devin.Lu@quantatw.com>
Change-Id: I7ded958e1d2722b2e48c21466ff6fd2f82ecc5e7