zephyr: fix false error regarding HOOKs priority
The check to verify that all HOOKs are run at the lowest task priority
was called before the the system work queue priority was adjusted. This
caused the following false positive error on the console:

    ERROR: sysworkq has priority -1 but must be >= 10

Move the check as the first HOOK_INIT, which is triggered after the EC
app corrects the system work queue priority.

BUG=none
BRANCH=none
TEST=zmake testall
TEST=boot zephyr on Herobrine, verify error message is no longer shown

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: I804dc17acffda5c21316939293aff5ab57bcba53