bq25710: Update IL_AVG handling
The original bq25710 datasheet specifies different IL_AVG encodings form
the bq25710 february 2021 revision. The update actually matches the
bq25720 encoding, so we can merge the two cases.

We only have one board family using the bq25720 and it calls for a 10 A
limit. We can leave the code as-is for now. There is no need for a
config option until another bq25720 design needs a different limit.

BRANCH=none
BUG=b:185190976
TEST=buildall passes

Signed-off-by: Caveh Jalali <caveh@chromium.org>
Change-Id: I2a7466dec4a2d35ffb33a6684d9c19600e627086