gingerbread: Add DP Rx EQ .board_set method for TUSB1064
This CL adds a .board_set method for the TUSB1064 usb_mux config to
control the display port rx EQ gain. Previously, this gain was being
applied in the driver init() function. The driver has been modified to
remove this default gain setting. So now the gain must be selected by
the boards that use the TUSB1064.

BUG=b:206059703
BRANCH=none
TEST=verified that DP Rx EQ gain is set when the usbc host port is
attached.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I4b8e3463aaa216aa1d71aa6a87b9819e6d78121a