Add EC feature flag for S4 residency
On systems with Intel Key Locker, the AP must hibernate to S4 in order
to preserve active Key Locker state. Until recently, the EC did not
support camping out in S4. Add a feature bit so the AP can distinguish
EC versions which are capable of going to S4 from those that are not.

BRANCH=None
BUG=b:204947672
TEST=hiberman hibernate --test-keys on volteer

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: I9d01c689e90cd479d8d3f06a224ef7f9ed309fc4