zephyr: gitlab: filter zephyr/drivers/** directories
Remove the drivers written under zephyr/drivers/ from the code coverage.
These drivers are in the process of being migrated upstream and the tests
will be done upstream as well. Additionally, adding emulators to support
this in a host test will be too difficult since many dependencies in
the Zephyr build are SOC and SOC_FAMILY dependent.

BRANCH=none
BUG=none
TEST=none

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: I27bcc31e5ee474aff2be050b8ef8289a1512b6cc