servo_v4p1: Add DP Rx EQ .board_set method for TUSB1064
This CL adds a .board_set method for the TUSB1064 usb_mux config to
control the display port rx EQ gain. Previously, this gain was being
applied in the driver init() function. The driver has been modified to
remove this default gain setting. So now the gain must be selected by
the boards that use the TUSB1064.

BUG=b:206059703
BRANCH=none
TEST=make -j BOARD=servo_v4p1

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I5e2ce514d11b65b23afb7e59805c4184cffa8e24