zephyr test: Use interrupt GPIO for SYV682x
Set and clear the ALERT_L signal in the SYV682x emulator based on the
state of underlying conditions. Test the driver by just setting the
underlying conditions instead of calling the ISR from within the test.
Define the interrupt signal in the device tree and enable interrupts
from it in the test.

BUG=b:190519131
TEST=zmake configure --test zephyr/test/drivers
BRANCH=none

Signed-off-by: Abe Levkoy <alevkoy@chromium.org>
Change-Id: Idc8b957fe53ce78bd8236e4af6b005636498ada2